Move types somewhere else
