+++ Cargo.toml
@@ -13,7 +13,6 @@ harness = false
 [dependencies]
 eyre = "0.6.8"
 futures = "0.3.24"
-httparse = "1.8.0"
 memmap2 = "0.5.7"
 nom = "7.1.1"
 smallvec = { version = "1.10.0", features = ["const_generics", "const_new", "union", "debugger_visualizer"] }
@@ -31,6 +30,7 @@ hyper = { version = "0.14.20", features = ["client", "server", "http1", "tcp", "
 pretty_assertions = "1.3.0"
 tokio-stream = "0.1.11"
 tracing-subscriber = "0.3.16"
+httparse = "1.8.0"
 
 [profile.dev.package."*"]
 opt-level = 2
